text2pdf/requirements.txt
requests
google-cloud-storage
google-cloud-pubsub
flask
fpdf
